+++ js/glotdict-settings.js
@@ -1,10 +1,4 @@
 jQuery('#menu-headline-nav').append('<li class="current-menu-item gd_setting" style="cursor:pointer;"><a style="font-weight:bold;"> GlotDict</a></li>');
-if (!gd_get_setting('hide_info_message')) {
-  if (jQuery(window).width() > 1390) {
-    jQuery('body').append('<div style="position: absolute;top: 320px;right: 10px;width:200px;background-color: #ddd;padding: .5em;">Don\’t forget to click on the <img class="gd_icon2">GlotDict icon in the blue menu!</div>');
-    jQuery('.gd_icon2').attr('src', jQuery('.gd_icon').attr('src'));
-  }
-}
 jQuery('.gd_icon').prependTo('.gd_setting').show();
 
 jQuery('.gd_setting').click(function() {
@@ -28,7 +22,7 @@ function gd_generate_settings_panel() {
     'no_non_breaking_space': 'Don’t visualize non-breaking-spaces in preview',
     'no_trailing_space': 'Hide warning for trailing space in translation',
     'curly_apostrophe_warning': 'Show warning for missing curly apostrophe in preview',
-    'hide_info_message': 'Hide Info messages about this menu'
+    'localized_quote_warning': 'Show warning for using non-typographic quotes in preview (except for HTML attributes quotes)'
   };
   var container = '<div class="notice gd_settings_panel"><h2>GlotDict Settings</h2></div>';
   jQuery('.gp-content').prepend(container);
@@ -43,6 +37,7 @@ function gd_generate_settings_panel() {
     '<li>Shortcut on Ctrl+Shift+B to "Copy from original"</li>' +
     '<li>Shortcut on Ctrl+Shift+F to add non-breaking spaces near symbols</li>' +
     '<li>Shortcut on Ctrl+Alt+R to reset all the GlotDict settings</li>' +
+    '<li>Shortcut on Ctrl+Alt+D to dismiss all the validation warnings</li>' +
     '<li>Right click of the mouse on the term with a dashed line and the translation will be added in the translation area</li>' +
     '</ul><br><h3>Settings</h3>';
   jQuery('.gd_settings_panel').append(hotkeys);