+++ js/glotdict.js
@@ -10,13 +10,15 @@ if (jQuery('.filters-toolbar:last div:first').length > 0) {
   if (jQuery('#bulk-actions-toolbar-top').length > 0) {
     jQuery('#upper-filters-toolbar').css('clear', 'both');
     gd_add_column();
-    jQuery('#bulk-actions-toolbar-top').clone().css('float', 'none').insertBefore('#legend');
-    jQuery('form.filters-toolbar.bulk-actions').submit(function() {
-      var row_ids = jQuery('input:checked', jQuery('table#translations th.checkbox')).map(function() {
-        return jQuery(this).parents('tr.preview').attr('row');
-      }).get().join(',');
-      jQuery('input[name="bulk[row-ids]"]', jQuery(this)).val(row_ids);
-    });
+    if (jQuery('#bulk-actions-toolbar-bottom').length === 0) {
+      jQuery('#bulk-actions-toolbar-top').clone().css('float', 'none').insertBefore('#legend');
+      jQuery('form.filters-toolbar.bulk-actions').submit(function() {
+        var row_ids = jQuery('input:checked', jQuery('table#translations th.checkbox')).map(function() {
+          return jQuery(this).parents('tr.preview').attr('row');
+        }).get().join(',');
+        jQuery('input[name="bulk[row-ids]"]', jQuery(this)).val(row_ids);
+      });
+    }
   }
   if (jQuery('.preview').length === 1) {
     jQuery('.preview .action').trigger('click');
@@ -43,6 +45,7 @@ if (jQuery('.filters-toolbar:last div:first').length > 0) {
 
 gd_add_project_links();
 gd_add_button();
+gd_add_meta();
 
 jQuery('.glotdict_language').change(function() {
   localStorage.setItem('gd_language', jQuery('.glotdict_language option:selected').text());
@@ -85,3 +88,5 @@ gd_non_breaking_space_highlight();
 gd_wait_table_alter();
 
 gd_remove_layover();
+
+jQuery('.gp-content').css('max-width', '85%');